Promote usage of isolated development environments
Using pythons virtualenv, virtualenvwrapper and Nodes nodeenv we can ensure that the node version, any of the required dependancies are installed into an isolated environment without modifying any system versions.

This has been the normal development routine in the python world. Ruby and Node have yet to perfect their tools for this. In the mean time, virtualenvwrapper greatly helps by keeping the node virtualenv out of the source code.